Extract ShardMerger specs into dedicated spec file
